# art/containers/model_device.go
/*
 * Container Lifecycle Service
 *
 * This files describes the container lifecycle service
 *
 * API version: 1.0.0
 * Contact: info@menucha.de
 */
package containers

type Device struct {
	Path string `json:"path,omitempty"`
}