# art/containers/model_namespace.go
/*
 * Container Lifecycle Service
 *
 * This files describes the container lifecycle service
 *
 * API version: 1.0.0
 * Contact: info@menucha.de
 */
package containers

type Namespace struct {
	Type_ string `json:"type,omitempty"`
}